I forgot this ...

+++ Module1.vb
@@ -448,9 +448,10 @@
                                 Loop While reset <> "Y" And reset <> "N"
 
                                 If reset = "Y" Then
-                                        My.Computer.FileSystem.DeleteDirectory("C:\Program Files\PocketMine-ManagerServers\Data", FileIO.DeleteDirectoryOption.DeleteAllContents)
+                                    If dirpath And dirservername And dirdata Then
                                         My.Computer.FileSystem.DeleteDirectory("C:\Program Files\PocketMine-ManagerServers\Path", FileIO.DeleteDirectoryOption.DeleteAllContents)
                                         My.Computer.FileSystem.DeleteDirectory("C:\Program Files\PocketMine-ManagerServers\ServersName", FileIO.DeleteDirectoryOption.DeleteAllContents)
+                                        My.Computer.FileSystem.DeleteDirectory("C:\Program Files\PocketMine-ManagerServers\Data", FileIO.DeleteDirectoryOption.DeleteAllContents)
                                         Console.WriteLine("Closing program...")
                                         End
                                     Else
